L10nTests-UseOnly2Clones
