+++ bitrise.yml
@@ -1766,7 +1766,7 @@ trigger_map:
 - push_branch: v32.1
   workflow: xcode12-release-and-beta-nocache
 - pull_request_target_branch: main
-  workflow: RunUnitTests
+  workflow: L10nBuild
 - pull_request_source_branch: '*'
   pull_request_target_branch: chronological-tabs
   workflow: RunUnitTests
